requirements.txt

streamlit

python-dotenv

google-generativeai

arxiv

unstructured[pdf]

PyMuPDF
pdfplumber
camelot-py[cv]

pandas

langchain
tiktoken

sentence-transformers
faiss-cpu

rank-bm25

numpy